pyproject.toml
[project]
name = "agentlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.5",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "langgraph-cli[inmem]>=0.3.3",
    "langsmith>=0.4.4",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
]
